custom ValidVariableNameSniff (renamed from CamelCaseParameterSniff) now provides functionality which is missing from default ValidVariableNameSniff
- method parameters should not start with underscore
- private class members are checked for camelCase format even when PrivateNoUnderscore property is disabled for default ValidVariableNameSniff
